+++ ChangeLog.md
@@ -17,6 +17,12 @@ Changelog for CEREScript-Core01
 Change type of `CERES` and `Value` as parametric to allow to change type of Str, etc.
 
 
+## 0.13.0.0 -- 2020-04-09
+
+### Changed
+* Add more `VariablePlace` and rename of `Local*`
+
+
 ## 0.12.0.0 -- 2020-04-06
 
 ### Changed